requirements.txt
streamlit
pandas
pandas_ta
python-binance
adjustText